Add ImageField that resizes/compresses the image
I have added a custom model field, ResizedHashNameImageField, that automatically resizes the uploaded image if its maximum dimension is larger than the limit set on the field. It will also save the image with a quality compression setting of 75. This will reduce the storage needs for uploaded images and optimize the loading times of the website.

In addition, this field also takes care of creating hash-based filenames.