Replace ImageFields by ResizedHashNameImageFields
I have replaced the ImageFields in the models that used it by the new ResizedHashedImageFields. While this does not affect images that have already been saved, this will make that new images will be automatically scaled and compressed on upload.

In addition, I've squashed the migration histories of the home and userprofile app. They contained a lot of "in progress" migrations that were no longer needed.