Add handling for png image format
The handling of images with an alpha layer, like some PNG images, was broken by forcing it to save in the JPEG format. Since replacing the alpha layer by a color would change the appearance of the image and PNG images benefit from the bitmap format, I've chosen to save images in their original format instead of forcing them to become JPEGs.